Edits to chapter 3.
I made a lot of small tweaks, but that's just because I'm being taking the time to really polish things in this cohort. These were great! Thank you!

Some particular things to note:
- Name chunks with CHAPTER-description, such as "03-basic-math". That way, if there's an error during the build, it's easier to find the broken code.
- Break things up into separate slides. You can use arrow keys to navigate while presenting, avoid having to scroll
- I started to do it here, but really more would be better: Use emojis to shorten obvious things, like ⬆️ instead of "up arrow".

(I'll repeat these notes in the review to make sure they get to you cleanly, sorry for the duplication if it's annoying!)