Condensed wording.
Don't use full sentences! I think a lot of this was already there, so I'm not being a stickler about making it perfect, but I edited the first several slides to show you the general idea. Remember that these are slides with discussion points, not a replacement for the book.