Cohort 10 chapter 07 notes (#112)
* Cohort 10 chapter 07 notes

* Condensed wording.

Don't use full sentences! I think a lot of this was already there, so I'm not being a stickler about making it perfect, but I edited the first several slides to show you the general idea. Remember that these are slides with discussion points, not a replacement for the book.

* Fix chapter 23 title.

---------

Co-authored-by: Jon Harmon <jonthegeek@gmail.com>